requirements.txt
supabase>=2.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.30.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
